refactor(report): 체결 품질 섹션을 ExecutionQualityReportBuilder 로 분리
strategy_log_report_service.py 가 2,716줄까지 커져 M-2 감시 항목에서 증가 가속으로
잡혔다. 리포트의 나머지와 상태를 공유하지 않는 체결 품질·후보 유동성 섹션(446줄)을
전용 빌더로 이관해 2,243줄로 내린다.

- services/strategy_execution_quality_report.py: 두 섹션 빌더 + 헬퍼(최신 레코드
  선별, 기간 라벨, 임계 판정). 비활성화 후보 목록은 빌더가 소유한다.
- services/strategy_report_format.py: 두 모듈이 함께 쓰는 포맷 헬퍼와 행 수 상수.
  서로 import 하지 않도록 단방향 의존(format ← service·builder)으로 만든다.
- 서비스의 get_last_execution_quality_candidates() 는 빌더로 위임만 하므로
  소비처(strategy_log_report_task)는 무변경.
- 테스트는 이관된 헬퍼 호출을 빌더 경유로 바꾼다(동작 검증 내용은 그대로).

단위 7,923 통과(기존 실패 1건은 샌드박스 egress 차단, baseline 동일),
통합 실패 목록 baseline 과 완전 동일.
