+++ todo_list.md
@@ -300,7 +300,8 @@
 - [x] `docs/backtest.md`(본문은 #619 슬리피지/스프레드 CLI까지 이미 반영, 날짜만 동기화)·`CODEBASE_SUMMARY.md`(bootstrap 분해 진전·EventShadowManager 분리·해외/테마 계층·브로커 계층 계측 4곳 반영) 갱신 완료 (2026-07-05).
 - [x] **`service_container.py` 비대화 경고 해소 — bootstrap 분해 + 의존성 경계 테스트 추가 (2026-07-12)**: #659(서비스↔전략 의존성 경계 강제, `oneil_common_types.py`를 `strategies/`→`common/`으로 이관, `test_architecture_dependencies.py` 신규) · #660(`backtest_task_bootstrap.py` 추출) · #661(`repository_bootstrap.py` 추출) · #662(`market_data_bootstrap.py` 추출) · #663(`query_bootstrap.py` 추출) · #664(`realtime_bootstrap.py` 추출) 6개 PR로 `service_container.py`가 962→**710줄**로 감소, 07-05까지의 증가 추세 반전.
 - [x] **해외 조립 분해 (2026-08-22)**: `service_container.py` 의 해외(미국장) 조립 4경로(관심종목 등락 알림 · dry-run 파이프라인 · 수동 주문 게이팅 · 장중 VBO 폴링, 220줄)를 `view/web/bootstrap/overseas_bootstrap.py`(`OverseasBootstrap`)로 이관했다 — **1,174→936줄**. #659~664 와 같은 bootstrap 추출 패턴이고 조립 순서·조건·후주입은 불변, 자동 경로 `live_enabled=False` 잠금과 수동 경로 `live_enabled=True` 의 인스턴스 분리도 그대로다(이제 이 파일이 유일한 조립 지점). 테스트는 patch 대상 모듈만 이관하되 양쪽 모듈이 공유하는 `USMarketCalendarService` 는 **`overseas_bootstrap.<이름>` 접두 키**로 등록했다 — 같은 키로 두 번 패치하면 나중 것이 앞의 것을 덮어써 어느 모듈의 mock 인지 알 수 없어지기 때문.
-- 감시(조치 아님, 2026-08-22 실측 — 07-18 이후 갱신되지 않아 실제 증가폭이 문서에 반영되지 않고 있었다): `scheduler/strategy_scheduler.py` 2,426→**2,499줄**(+73, 08-18 이후 정체) / `services/strategy_log_report_service.py` 2,298→**2,716줄**(+418, 08-18 2,549 → 08-21 2,716 로 증가 가속) / `view/web/bootstrap/service_container.py` 746→1,174줄(+428; 08-18 1,078 → 08-21 1,126)까지 늘어 **#659~664 분해로 얻은 962→710 감소가 상쇄되고 분해 이전(962줄)보다 커졌다가**, 위 해외 분해로 **936줄**로 내렸다. 다만 신규 서비스 조립이 계속 이 파일에 쌓이는 구조 자체는 그대로라 분해만으로는 재발을 막지 못한다 — 다음 증가 시엔 조립 지점 규약(신규 서비스군은 전용 bootstrap 모듈에 넣는다)을 함께 정할 것. 남은 god class 분해는 3-4 재승격과 함께 진행.
+- [x] **체결 품질 섹션 분리 (2026-08-22)**: `strategy_log_report_service.py` 의 체결 품질·후보 유동성 섹션과 그 헬퍼(최신 레코드 선별·기간 라벨·임계 판정, 446줄)를 `services/strategy_execution_quality_report.py`(`ExecutionQualityReportBuilder`)로 이관했다 — **2,716→2,243줄**. 두 모듈이 함께 쓰는 포맷 헬퍼(`_esc`/`_to_float`/`_format_eok_won`/`_first_number`)와 행 수 상수는 `services/strategy_report_format.py` 로 빼 단방향 의존(format ← service·builder)을 만들었다. 비활성화 후보 상태는 빌더가 소유하되 서비스의 `get_last_execution_quality_candidates()` 는 그대로라 소비처(`strategy_log_report_task`)는 무변경.
+- 감시(조치 아님, 2026-08-22 실측 — 07-18 이후 갱신되지 않아 실제 증가폭이 문서에 반영되지 않고 있었다): `scheduler/strategy_scheduler.py` 2,426→**2,499줄**(+73, 08-18 이후 정체) / `services/strategy_log_report_service.py` 2,298→2,716줄(+418, 08-18 2,549 → 08-21 2,716 로 증가 가속) 후 위 분리로 **2,243줄** / `view/web/bootstrap/service_container.py` 746→1,174줄(+428; 08-18 1,078 → 08-21 1,126)까지 늘어 **#659~664 분해로 얻은 962→710 감소가 상쇄되고 분해 이전(962줄)보다 커졌다가**, 위 해외 분해로 **936줄**로 내렸다. 다만 신규 서비스 조립이 계속 이 파일에 쌓이는 구조 자체는 그대로라 분해만으로는 재발을 막지 못한다 — 다음 증가 시엔 조립 지점 규약(신규 서비스군은 전용 bootstrap 모듈에 넣는다)을 함께 정할 것. 남은 god class 분해는 3-4 재승격과 함께 진행.
 
 ### M-3. 시가총액/미국주식 UI 하드닝 후속 [#653 코드리뷰 xhigh, 2026-07-11]
